requirements.txt
streamlit
groq
pdfplumber
python-docx
python-dotenv
